Extract `Layer`s from `sine` script
