+++ mlp/__init__.py
@@ -1,0 +1,11 @@
+# coding: utf-8
+
+from __future__ import (
+    absolute_import,
+    division,
+    print_function,
+    unicode_literals,
+)
+
+
+from .mlp import *
